refactor: rename nimUtils.ts to legacyNimUtils.ts
Update import paths in legacyModelServingNim.cy.ts and
modelServingNim.cy.ts to match the renamed file.
